api/requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Authentication
Flask-JWT-Extended==4.6.0
werkzeug==3.0.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Validation
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Environment
python-dotenv==1.0.0

# API Documentation
flask-swagger-ui==4.11.1
PyYAML==6.0.1

# Password hashing (already in werkzeug, but explicit for clarity)
# bcrypt==4.1.1  # Alternative if needed

# Development
pytest==7.4.3
pytest-flask==1.3.0
